requirements.txt
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.12
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.11.0
dataclasses-json==0.6.0
datasets==2.14.5
dill==0.3.7
fastapi==0.99.1
ffmpy==0.3.1
filelock==3.12.4
FlagEmbedding==1.1.1
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gradio==3.44.4
gradio_client==0.5.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.299
langsmith==0.0.40
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
onnxruntime==1.16.0
orjson==3.9.7
overrides==7.4.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
playwright==1.38.0
posthog==3.0.2
protobuf==4.24.3
psutil==5.9.5
pulsar-client==3.3.0
pyarrow==13.0.0
pydantic==1.10.12
pydub==0.25.1
pyee==9.0.4
pyparsing==3.1.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sseclient-py==1.7.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
together==0.1.9
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
xxhash==3.3.0
yarl==1.9.2
